80% Done

+++ checkout/delete.php
@@ -1,16 +1,19 @@
 <?php // sql to delete a record
+ini_set('display_errors', 1);
+ini_set('display_startup_errors', 1);
+error_reporting(E_ALL);
 $conn = new PDO('mysql:host=localhost;dbname=pescasamac', 'root', '');
 if (!$conn) {
     die("Fatal Error: Connection Failed!");
 }
 
-
-$query = $conn->prepare("DELETE from `cart`");
-$query->execute();
-header("location:checkout.php");
+if (isset($_POST['deleteAll'])) {
+    $query = $conn->prepare("DELETE from `cart`");
+    $query->execute();
+    header("location:checkout.php");
+}
 ?>
 <script>
     alertify.set('notifier', 'position', 'top-right');
-    alertify.success('You delete a;; pf your products from Cart!');
-    location.replace("checkout.php");
+    alertify.success('You delete all of your products from Cart!');
 </script>
